Frontend: Consolidate parser contract tests in frontend suites
